+++ index.js
@@ -1,16 +1,19 @@
 const Logger = require('./src/Transports/index.js');
 
+/* 
 const castoeLog = new Logger.File({
 	file: 'file.txt',
 	automatic: true,
-	overwrite: true
+	overwrite: true,
 });
-
-castoeLog.send('Testing');
+castoeLog.send('Testing'); */
 
 const castoeConsole = new Logger.Console({
-	name: 'Castoe COnsole',
-	date: 'LT'	
-});
+	name: 'Castoe Console',
+	date: 'LT',
+	showType: true
+})
 
-castoeConsole.send('Test');
+const x = true;
+
+castoeConsole.send(x);